Implement SPDX 3 hasConcludedLicense
