+++ ntia_conformance_checker/ntia_checker.py
@@ -39,9 +39,6 @@ class NTIAChecker(BaseChecker):
             raise ValueError("Only NTIA Minimum Element compliance is supported.")
 
         if self.doc:
-            self.doc_author = True  # Assume author is present?
-            self.doc_timestamp = True  # Assume timestamp is present?
-            self.dependency_relationships = self.check_dependency_relationships()
             self.compliant = self.check_compliance()
 
             # for backward compatibility
